Moved dashboard permissions into components dir
